pyproject.toml
[tool.poetry]
name = "django-vite-example"
version = "1.0.0"
description = "A example Django project with Django Vite"
authors = ["MrBin99"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.0.0"
django-vite = "^2.0.0"
inertia-django = "^0.3.2"
django-js-routes = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
